requirements.txt
Django==2.2.28
django-filter==2.2.0
djangorestframework==3.11.2
Markdown==3.1.1
pytz==2022.1
sqlparse==0.4.2
